PostgreSQL and Java project for the course "Databases"
This is my first Database implementation in PostgreSQL (and SQL in
general). The assignment was offered by the course "Databases" of ECE,
TUC, Greece in Spring Semester 2017.

Phase A: The description, requirements and the ER diagram of a
university department database are given. Students were asked to
implement the corresponding relational database with the appropriate
tables, functions, triggers and views.

Phase B: Having completed the database from Phase A, we continued with
performance optimization of specific queries. At this state we studied
initial query plans and added appropriate indices, clustering,
materialized views to improve time efficiency. Furthermore, a sample
java app was connected with the database using JDBC.

Files in this repository include:
1) Backup of PostgreSQL database from Phase B
2) A Java project using JDBC
3) Final report explaining the use of indices and clustering for search
optimization (in Greek)
4) Assignment description files of Phase A and Phase B (in Greek)